+++ src/schemas/index.ts
@@ -1,3 +1,4 @@
+export { optionalEmail, taxTypeSchema } from './common'
 export { categorySchema, type CategoryFormValues } from './category.schema'
 export { productSchema, type ProductFormValues } from './product.schema'
 export { supplierSchema, type SupplierFormValues } from './supplier.schema'
