+++ src/utils/dateToIsoStartOfDay.ts
@@ -1,0 +1,4 @@
+export function dateToIsoStartOfDay(value?: string | null): string | undefined {
+    if (!value) return undefined
+    return `${value}T00:00:00Z`
+}
